riskAssessment/requirements.txt
Flask==2.1.2
numpy==1.21.0
scikit-learn==1.0.2
pandas==1.3.5
neptune-client==0.16.3
python-dotenv==0.20.0